# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InvitationAccept": {
        "description": "Schema for accepting an invitation.",
        "properties": {
          "password": {
            "description": "Password for the new account (min 6 characters)",
            "minLength": 6,
            "title": "Password",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "description": "User's timezone preference",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "InvitationAccept",
        "type": "object"
      },
      "InvitationAcceptResponse": {
        "description": "Schema for invitation acceptance response.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "person_id": {
            "title": "Person Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "person_id",
          "org_id",
          "name",
          "email",
          "roles",
          "timezone",
          "token",
          "message"
        ],
        "title": "InvitationAcceptResponse",
        "type": "object"
      },
      "InvitationCreate": {
        "description": "Schema for creating an invitation.",
        "properties": {
          "email": {
            "description": "Email address of the invitee",
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "description": "Full name of the invitee",
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "roles": {
            "description": "Roles to assign (e.g., ['volunteer', 'admin'])",
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "email",
          "name",
          "roles"
        ],
        "title": "InvitationCreate",
        "type": "object"
      },
      "InvitationResponse": {
        "description": "Schema for invitation response.",
        "properties": {
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "invited_by": {
            "title": "Invited By",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "org_id",
          "email",
          "name",
          "roles",
          "invited_by",
          "token",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "InvitationResponse",
        "type": "object"
      },
      "InvitationVerify": {
        "description": "Schema for verifying an invitation token.",
        "properties": {
          "invitation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvitationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "InvitationVerify",
        "type": "object"
      },
...
      "ListResponse_InvitationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvitationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ListResponse[InvitationResponse]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered volunteer scheduling and sign-up management",
    "title": "SignUpFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/invitations": {
      "get": {
        "description": "List all invitations for an organization (admin only).\n\nRequires JWT authentication via Authorization header.",
        "operationId": "listInvitations",
        "parameters": [
          {
            "description": "Organization ID",
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "description": "Organization ID",
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "description": "Filter by status (pending, accepted, expired, cancelled)",
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (pending, accepted, expired, cancelled)",
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_InvitationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Invitations",
        "tags": [
          "invitations"
        ]
      },
      "post": {
        "description": "Create a new invitation (admin only). Rate limited to 10 requests per 5 minutes per IP.\n\nSends an invitation email to a new user to join the organization.",
        "operationId": "createInvitation",
        "parameters": [
          {
            "description": "Organization ID",
            "in": "query",
            "name": "org_id",
            "required": true,
            "schema": {
              "description": "Organization ID",
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{invitation_id}": {
      "delete": {
        "description": "Cancel a pending invitation (admin only).",
        "operationId": "cancelInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{invitation_id}/resend": {
      "post": {
        "description": "Resend an invitation email (admin only).\n\nGenerates a new token and extends the expiry date.",
        "operationId": "resendInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "title": "Invitation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resend Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{token}": {
      "get": {
        "description": "Verify an invitation token. Rate limited to 10 requests per minute per IP.\n\nChecks if the invitation is valid and not expired.",
        "operationId": "verifyInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationVerify"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{token}/accept": {
      "post": {
        "description": "Accept an invitation and create a new account.\n\nThis creates a new Person with the invited roles and marks the invitation as accepted.",
        "operationId": "acceptInvitation",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationAcceptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
  }
}